% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REG.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot Method for linreg Objects}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{linreg}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns a list of ggplot objects.
}
\description{
Creates diagnostic plots: residuals vs fitted and scale-location.
}
