% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REG.R
\name{pred}
\alias{pred}
\title{Prediction Generic Function}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{A model object (can be any object for which a method has been defined).}

\item{...}{Additional arguments passed to methods.}
}
\description{
A generic function for obtaining predictions from model objects.
}
