% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REG.R
\name{print.linreg}
\alias{print.linreg}
\title{Print Method for linreg Objects}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{linreg}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Displays the formula and estimated coefficients.
}
