% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REG.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Residuals Method for linreg Objects}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A numeric vector of residuals.
}
\description{
Returns the residuals from the fitted model.
}
