% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REG.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary Method for linreg Objects}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Displays a table of coefficients, standard errors, t-values, and p-values.
}
